Refactor redirect routes
